Add ZQuery support and split example
